Records management - specific definition
Lines 620-21 (in PDF): “Ensure that records management programs provide adequate and proper documentation of agency activities.”

A specific definition needs to be given for “adequate and proper.”  The current language is too vague and open-ended.  We suggest language that reflects the roles that records management is required to serve, in line with “Records help us claim our rights and entitlements, hold our elected officials accountable for their actions, and document our history as a nation. In short, NARA ensures continuing access to the essential documentation of the rights of American citizens and the actions of their Government.” 

Reference: http://www.archives.gov/publications/general-info-leaflets/1-about-archives.html