Modify and clarify language
Lines 659-62 (in PDF): ”b) Ensure that the public has timely and equitable online access to the agency’s public information using a manner that is informed directly by public engagement and balanced against the costs of dissemination or accessibility improvements and demonstrate usefulness of the information.”

This language requires modification. Once federal information is digital, the costs of dissemination are minimal, and accessibility improvements should be taken into consideration as a system or application is procured or developed. Clarification is also needed as to whether “accessibility” here is intended to refer to Sec. 508 requirements.  The language after ‘improvements’ should be struck; it is a remnant of arguments about whether the government should be in the role of providing information to the public – which has been settled.