Restore the following language from the 2000 Circular:
Following line 216 (of the PDF)

Restore the following language from the November 28, 2000 Revision of the Circular: 

“n. Users of Federal information resources must have skills, knowledge, and training to manage information resources, enabling the Federal government to effectively serve the public through automated means.”
o. The application of up-to-date information technology presents opportunities to promote fundamental changes in agency structures, work processes, and ways of interacting with the public that improve the effectiveness and efficiency of Federal agencies.” 
p. The availability of government information in diverse media, including electronic formats, permits agencies and the public greater flexibility in using the information.”
q. Federal managers with program delivery responsibilities should recognize the importance of information resources management to mission performance.”
r. The Chief Information Officers Council and the Information Technology Resources Board will help in the development and operation of interagency and interoperable shared information resources to support the performance of government missions.”