New language on FOIA guidance
Information Management and Access

After line 640 (in PDF): New language should be added to elaborate on the minimum standards on proactive disclosure and to ensure that agencies follow the presumption of disclosure, in accordance with the President’s 2009 and the Attorney General FOIA guidance.

“i. Agencies are required to proactively disclosure information when the agency official responsible for the information does not reasonably foresee that disclosure would harm an interest protected by an exemption described in subsection (b) of the Freedom of Information Act (5 U.S.C § 552). 

ii. Information and documents that are frequently requested by the public should be proactively released, including documents relating to health, safety and environmental issues. 

iii. Agencies are required to adopt a presumption of disclosure and openness, by evaluating records with a view towards what can be disclosed, rather than what can be withheld. Agencies shall not withhold information simply because existing statutes allow the withholding of such information.”